feat: update Terms of Service page layout and styling, add header and footer
feat: enhance Dialog component with center alignment and improved styles

chore: add ESLint configuration for Next.js and TypeScript

feat: create LoginPanel component for user authentication with Google and GitHub

feat: implement PrivacyPanel component to display privacy policy

feat: develop ProfilePanel component for user profile management and journal preferences

feat: create TermsPanel component to display terms of service

feat: implement UpvotedPanel component to manage and display upvoted articles
