+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@auth/prisma-adapter": "^2.10.0",
-    
     "@hookform/resolvers": "^4.1.3",
     "@radix-ui/react-accordion": "^1.2.3",
     "@radix-ui/react-alert-dialog": "^1.1.6",
@@ -42,7 +41,6 @@
     "date-fns": "^3.6.0",
     "fast-xml-parser": "^5.2.5",
     "firebase": "^11.3.0",
-    
     "lucide-react": "^0.475.0",
     "mongodb": "^7.2.0",
     "next": "^15.5.18",
@@ -64,6 +62,8 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "eslint": "9.39.4",
+    "eslint-config-next": "16.2.6",
     "postcss": "^8.5.13",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
